V2025.3.1

+++ include/system/hardwareinfo.h
@@ -12,12 +12,12 @@ namespace Nickvision::System::HardwareInfo
      * @brief Gets the amount of RAM in the system in bytes.
      * @return The amount of RAM in bytes
      */
-    unsigned long getTotalRamSize();
+    unsigned long long getTotalRamSize();
     /**
      * @brief Gets the amount of RAM free in the system in bytes.
      * @return The amount of free RAM in bytes
      */
-    unsigned long getFreeRamSize();
+    unsigned long long getFreeRamSize();
 }
 
 #endif //HARDWAREINFO_H
